Develop

+++ src/main/java/omsu/imit/schedule/service/ClassroomService.kt
@@ -2,37 +2,48 @@ package omsu.imit.schedule.service
 
 import omsu.imit.schedule.dto.request.CreateClassroomRequest
 import omsu.imit.schedule.dto.request.EditClassroomRequest
-import omsu.imit.schedule.dto.response.ClassroomShortInfo
-import omsu.imit.schedule.dto.response.ClassroomsByBuildingInfo
-import omsu.imit.schedule.dto.response.MetaInfo
+import omsu.imit.schedule.dto.response.*
+import omsu.imit.schedule.exception.CommonValidationException
 import omsu.imit.schedule.exception.ErrorCode
 import omsu.imit.schedule.exception.NotFoundException
 import omsu.imit.schedule.model.Building
 import omsu.imit.schedule.model.Classroom
+import omsu.imit.schedule.model.Day
+import omsu.imit.schedule.model.EventPeriod
 import omsu.imit.schedule.repository.ClassroomRepository
+import omsu.imit.schedule.repository.EventPeriodRepository
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.dao.DataIntegrityViolationException
 import org.springframework.data.domain.PageRequest
 import org.springframework.data.domain.Pageable
 import org.springframework.data.domain.Sort
 import org.springframework.stereotype.Service
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
+import java.util.*
 import kotlin.math.roundToInt
 
 @Service
 class ClassroomService
 @Autowired
-constructor(private val classroomRepository: ClassroomRepository,
-            private val buildingService: BuildingService,
+constructor(private val buildingService: BuildingService,
+            private val classroomRepository: ClassroomRepository,
+            private val eventPeriodRepository: EventPeriodRepository,
             private val tagService: TagService) : BaseService() {
 
     fun createClassroom(request: CreateClassroomRequest): ClassroomShortInfo {
         val building = buildingService.getBuildingById(request.buildingId);
-        if (classroomRepository.findByBuildingAndNumber(request.buildingId, request.number) != null)
-            throw NotFoundException(ErrorCode.CLASSROOM_ALREADY_EXISTS, request.number, request.buildingId.toString())
-
         val classroom = Classroom(building, request.number)
-        if (!request.tags.isNullOrEmpty()) classroom.tags = tagService.getAllTagsByIds(request.tags!!)
 
-        classroomRepository.save(classroom)
+        if (!request.tags.isNullOrEmpty())
+            classroom.tags = tagService.getAllTagsByIds(request.tags!!)
+
+        try {
+            classroomRepository.save(classroom)
+        } catch (e: DataIntegrityViolationException) {
+            throw CommonValidationException(ErrorCode.CLASSROOM_ALREADY_EXISTS, request.number, request.buildingId.toString())
+        }
+
         return toClassroomShortInfo(classroom)
     }
 
@@ -42,15 +53,11 @@ constructor(private val classroomRepository: ClassroomRepository,
                 .orElseThrow { NotFoundException(ErrorCode.CLASSROOM_NOT_EXISTS, classroomId.toString()) }
     }
 
-    fun getClassroomsByTags(tags: List<String>): Any {
+    fun getClassroomsByTags(tags: List<String>): List<ClassroomShortInfo> {
         return classroomRepository.findAllByTags(tags).asSequence().map { toClassroomShortInfo(it) }.toList();
     }
 
-    fun getAllClassroomssByBuilding(buildingId: Int): List<Classroom>? {
-        return classroomRepository.findAllByBuilding(buildingId, Sort.by("number"))
-    }
-
-    fun getAllClassroomssByBuilding(buildingId: Int, page: Int, size: Int): ClassroomsByBuildingInfo {
+    fun getAllClassroomsByBuilding(buildingId: Int, page: Int, size: Int): ClassroomsByBuildingInfo {
         val building: Building = buildingService.getBuildingById(buildingId);
         val pageable: Pageable = PageRequest.of(page, size, Sort.by("number"))
 
@@ -60,6 +67,16 @@ constructor(private val classroomRepository: ClassroomRepository,
                 .map { toClassroomShortInfo(it) }
                 .toList()
         return ClassroomsByBuildingInfo(createMetaInfo(building, page, size), classrooms)
+    }
+
+    fun getClassroomWithEventsByDate(classroomId: Int, searchDate: Date): ClassroomInfoByDate {
+        val classroom = getClassroomById(classroomId)
+        val date = LocalDate.parse(searchDate.toString(), DateTimeFormatter.ofPattern("yyyy-MM-dd"))
+        val day = Day.valueOf(date.dayOfWeek.toString())
+        val events = eventPeriodRepository.findAllByClassroomDayAndDate(classroomId, day, searchDate)
+        val response = createClassroomInfoByDate(searchDate, classroom, events)
+        println(response)
+        return response
     }
 
     fun editClassroom(classroomId: Int, request: EditClassroomRequest): ClassroomShortInfo {
@@ -74,6 +91,9 @@ constructor(private val classroomRepository: ClassroomRepository,
     }
 
     fun deleteClassroom(classroomId: Int) {
+        if (!classroomRepository.existsById(classroomId))
+            throw NotFoundException(ErrorCode.CLASSROOM_NOT_EXISTS, classroomId.toString())
+
         classroomRepository.deleteById(classroomId)
     }
 
@@ -84,27 +104,45 @@ constructor(private val classroomRepository: ClassroomRepository,
     private fun createMetaInfo(building: Building, page: Int, size: Int): MetaInfo {
         val baseUrl = "api/buildings/"
         val total = classroomRepository.countClassroomssByBuilding(building).toInt()
+        val lastPageNumber = ((total / size) - 1).toFloat().roundToInt()
 
-        var next: String? = null
-        var prev: String? = null
-        val first = baseUrl + "${building.id}/classrooms?page=0&size=$size"
-        val last = baseUrl + building.id + "?page=" + (total / size).toFloat().roundToInt() + "&size=" + size
+        var nextPage: String? = null
+        var prevPage: String? = null
+        val firstPage = "${baseUrl}${building.id}/classrooms?page=0&size=${size}"
+        val lastPage = "${baseUrl}${building.id}?page=${lastPageNumber}&size=${size}"
 
 
-        if (Math.round((total / size).toFloat()) > page) {
-            next = baseUrl + building.id + "/classrooms?page=" + (page + 1) + "&size=" + size
+        if (lastPageNumber > page) {
+            nextPage = "${baseUrl}${building.id}/classrooms?page=${page + 1}&size=${size}"
         }
         if (page != 0) {
-            prev = baseUrl + building.id + "/classrooms?page=" + (page - 1) + "&size=" + size
+            prevPage = "${baseUrl}${building.id}/classrooms?page=${page - 1}&size=${size}"
         }
 
         return MetaInfo(
                 total,
                 page,
                 size,
-                next,
-                prev,
-                first,
-                last)
+                nextPage,
+                prevPage,
+                firstPage,
+                lastPage)
+    }
+
+    private fun createClassroomInfoByDate(date: Date, classroom: Classroom, eventPeriods: List<EventPeriod>): ClassroomInfoByDate {
+        val events: MutableMap<Int, EventInfo> = mutableMapOf()
+
+        eventPeriods.asSequence().forEach { eventPeriod ->
+            if (!events.containsKey(eventPeriod.event.id)) {
+                events[eventPeriod.event.id] = toEventShortInfo(eventPeriod.event)
+            }
+            events[eventPeriod.event.id]!!.eventPeriods.add(toEventPeriodShortInfo(eventPeriod))
+        }
+
+        println(events.values)
+        return ClassroomInfoByDate(
+                date,
+                toClassroomShortInfo(classroom),
+                events.values)
     }
 }
